Support devices with amber/green LEDs
On devices like the Legend HTC used a binary amber/green LED with
support for hardware blinking.
Both colors on at the same time result in a slightly lighter amber that
is not distinguishable from amber at all. So we have to make sure their
status is mutual exclusive.

Amber is used as charging and battery low indicator and green for
notifications. Green always overrides amber.

Conflicts:
	core/java/android/provider/Settings.java
	core/res/res/values/config.xml
	services/java/com/android/server/NotificationManagerService.java

Change-Id: Ieafa68bd8fc75cd8836c3f4b06972de09c5f21d5