Fix notification airplane mode toggle to send the proper state, in turn fixing toggling of 4G when entering/leaving airplane mode.
The state in the broadcast intent was backwards.
It didn't affect the toggling of other radios, because they ignore the
intent payload and check the system setting.  4G toggling does use the
state extra, so it was turning 4G on when airplane mode was enabled,
and 4G off when airplane mode was disabled via the notification widget.

Change-Id: Ifda805136c08b7f427d44f539843a50512d3bf20