Try to properly create a bitmap from a drawable
Tries to create a bitmap from a drawable object using a canvas that
draws onto the new bitmap, which is the passed to the drawable
which draws onto the canvas, which is drawing onto the bitmap.
The final result should be a bitmap containing the drawable image.

Cherry-pick these commits:
https://github.com/LiquidSmoothROMs/android_frameworks_base/commit/9a233952a56a512d03e18452f8e1c6bc77385dee
https://github.com/LiquidSmoothROMs/android_frameworks_base/commit/314c9ac8df72468a5761ad558065a861166adc51
https://github.com/LiquidSmoothROMs/android_frameworks_base/commit/dcc5713effa5e0c9102c55f4bf34c8df20a8d426